requirements.txt
Flask==3.0.3
boto3==1.34.16
gunicorn==20.1.0
Pillow==10.0.0
google-generativeai==0.7.2